us-public-server/requirements.txt
# Web框架
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# 数据库
sqlalchemy==2.0.36

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1

# HTTP客户端
requests==2.32.3
httpx==0.28.1

# RSOE抓取
beautifulsoup4==4.12.3
lxml==5.3.0

# Google Earth Engine
earthengine-api==1.3.0

# OpenAI（质量评估）
openai==1.59.3

# Google Gemini（日报生成）
google-generativeai==0.8.3

# 任务调度
apscheduler==3.10.4

# 日志
loguru==0.7.3

# 工具
pydantic==2.10.4
